add operator shuffle build (#16145)
把shuffle join和broadcast join的build逻辑独立出来，为后续的优化以及重构做准备

Approved by: @m-schen, @aunjgr, @XuPeng-SH

+++ pkg/sql/colexec/external/external.go
@@ -180,7 +180,7 @@ func (arg *Argument) Call(proc *process.Process) (vm.CallResult, error) {
 	}
 	result.Batch = arg.buf
 	if result.Batch != nil {
-		result.Batch.ShuffleIDX = param.Idx
+		result.Batch.ShuffleIDX = int32(param.Idx)
 	}
 	return result, nil
 }
